+++ tap_doubler.c
@@ -84,35 +84,35 @@ typedef struct {
 	LADSPA_Data old_pitch;
 
 	LADSPA_Data * ring_L;
-	unsigned long buflen_L;
-	unsigned long pos_L;
+	int buflen_L;
+	int pos_L;
 
 	LADSPA_Data * ring_R;
-	unsigned long buflen_R;
-	unsigned long pos_R;
+	int buflen_R;
+	int pos_R;
 
 	LADSPA_Data * ring_pnoise;
-	unsigned long buflen_pnoise;
-	unsigned long pos_pnoise;
+	int buflen_pnoise;
+	int pos_pnoise;
 
 	LADSPA_Data * ring_dnoise;
-	unsigned long buflen_dnoise;
-	unsigned long pos_dnoise;
+	int buflen_dnoise;
+	int pos_dnoise;
 
 	float delay;
 	float d_delay;
 	float p_delay;
-	unsigned long n_delay;
+	int n_delay;
 
 	float pitchmod;
 	float d_pitch;
 	float p_pitch;
-	unsigned long n_pitch;
+	int n_pitch;
 
-	unsigned long p_stretch;
-	unsigned long d_stretch;
+	int p_stretch;
+	int d_stretch;
 
-	unsigned long sample_rate;
+	int sample_rate;
 	LADSPA_Data run_adding_gain;
 } Doubler;
 
@@ -203,7 +203,7 @@ void
 activate_Doubler(LADSPA_Handle Instance) {
 
 	Doubler * ptr = (Doubler *)Instance;
-	unsigned long i;
+	int i;
 
 	for (i = 0; i < BUFLEN * ptr->sample_rate / 192000; i++) {
 		ptr->ring_L[i] = 0.0f;
@@ -289,8 +289,8 @@ run_Doubler(LADSPA_Handle Instance,
 	LADSPA_Data * output_L = ptr->output_L;
 	LADSPA_Data * output_R = ptr->output_R;
 
-	unsigned long sample_index;
-	unsigned long sample_count = SampleCount;
+	int sample_index;
+	int sample_count = SampleCount;
 
 	LADSPA_Data in_L = 0.0f;
 	LADSPA_Data in_R = 0.0f;
@@ -381,14 +381,14 @@ run_Doubler(LADSPA_Handle Instance,
 		rem = fpos - n;
 
 		s_a_L = read_buffer(ptr->ring_L, ptr->buflen_L,
-				    ptr->pos_L, (unsigned long) n);
+				    ptr->pos_L, (int) n);
 		s_b_L = read_buffer(ptr->ring_L, ptr->buflen_L,
-				    ptr->pos_L, (unsigned long) n + 1);
+				    ptr->pos_L, (int) n + 1);
 
 		s_a_R = read_buffer(ptr->ring_R, ptr->buflen_R,
-				    ptr->pos_R, (unsigned long) n);
+				    ptr->pos_R, (int) n);
 		s_b_R = read_buffer(ptr->ring_R, ptr->buflen_R,
-				    ptr->pos_R, (unsigned long) n + 1);
+				    ptr->pos_R, (int) n + 1);
 
 		drystream_L = drylevel * in_L;
 		drystream_R = drylevel * in_R;
@@ -438,8 +438,8 @@ run_adding_Doubler(LADSPA_Handle Instance,
 	LADSPA_Data * output_L = ptr->output_L;
 	LADSPA_Data * output_R = ptr->output_R;
 
-	unsigned long sample_index;
-	unsigned long sample_count = SampleCount;
+	int sample_index;
+	int sample_count = SampleCount;
 
 	LADSPA_Data in_L = 0.0f;
 	LADSPA_Data in_R = 0.0f;
@@ -530,14 +530,14 @@ run_adding_Doubler(LADSPA_Handle Instance,
 		rem = fpos - n;
 
 		s_a_L = read_buffer(ptr->ring_L, ptr->buflen_L,
-				    ptr->pos_L, (unsigned long) n);
+				    ptr->pos_L, (int) n);
 		s_b_L = read_buffer(ptr->ring_L, ptr->buflen_L,
-				    ptr->pos_L, (unsigned long) n + 1);
+				    ptr->pos_L, (int) n + 1);
 
 		s_a_R = read_buffer(ptr->ring_R, ptr->buflen_R,
-				    ptr->pos_R, (unsigned long) n);
+				    ptr->pos_R, (int) n);
 		s_b_R = read_buffer(ptr->ring_R, ptr->buflen_R,
-				    ptr->pos_R, (unsigned long) n + 1);
+				    ptr->pos_R, (int) n + 1);
 
 		drystream_L = drylevel * in_L;
 		drystream_R = drylevel * in_R;
@@ -701,7 +701,7 @@ _init() {
 
 void
 delete_descriptor(LADSPA_Descriptor * descriptor) {
-	unsigned long index;
+	int index;
 	if (descriptor) {
 		free((char *)descriptor->Label);
 		free((char *)descriptor->Name);
