+++ tap_eq.c
@@ -301,7 +301,7 @@ run_eq(LADSPA_Handle instance, unsigned long sample_count) {
 	biquad * filters = ptr->filters;
 	float fs = ptr->fs;
 
-	unsigned long pos;
+	int pos;
 	float samp;
 
 
@@ -417,7 +417,7 @@ run_adding_eq(LADSPA_Handle instance, unsigned long sample_count) {
 	biquad * filters = ptr->filters;
 	float fs = ptr->fs;
 
-	unsigned long pos;
+	int pos;
 	float samp;
