+++ tap_pinknoise.c
@@ -56,10 +56,10 @@ typedef struct {
 	LADSPA_Data * output;
 
 	LADSPA_Data * ring;
-	unsigned long buflen;
-	unsigned long pos;
+	int buflen;
+	int pos;
 
-	unsigned long sample_rate;
+	int sample_rate;
 	LADSPA_Data run_adding_gain;
 } Pinknoise;
 
@@ -158,7 +158,7 @@ run_Pinknoise(LADSPA_Handle Instance,
 	LADSPA_Data hurst = LIMIT(*(ptr->hurst), 0.0f, 1.0f);
 	LADSPA_Data signal = db2lin(LIMIT(*(ptr->signal), -90.0f, 20.0f));
 	LADSPA_Data noise = db2lin(LIMIT(*(ptr->noise), -90.0f, 20.0f));
-	unsigned long sample_index;
+	int sample_index;
 	
   	for (sample_index = 0; sample_index < SampleCount; sample_index++) {
 
@@ -195,7 +195,7 @@ run_adding_Pinknoise(LADSPA_Handle Instance,
 	LADSPA_Data hurst = LIMIT(*(ptr->hurst), 0.0f, 1.0f);
 	LADSPA_Data signal = db2lin(LIMIT(*(ptr->signal), -90.0f, 20.0f));
 	LADSPA_Data noise = db2lin(LIMIT(*(ptr->noise), -90.0f, 20.0f));
-	unsigned long sample_index;
+	int sample_index;
 	
   	for (sample_index = 0; sample_index < SampleCount; sample_index++) {
 
@@ -309,7 +309,7 @@ _init() {
 
 void
 delete_descriptor(LADSPA_Descriptor * descriptor) {
-	unsigned long index;
+	int index;
 	if (descriptor) {
 		free((char *)descriptor->Label);
 		free((char *)descriptor->Name);
